+++ R/render_site.R
@@ -91,6 +91,11 @@
 #' \code{MASS::filter}. The default behaviour of \code{render_site} is to use a
 #' common R session.
 #'
+#' \code{autospin: true} causes \code{.R} files to be spinned and rendered
+#' (as well as \code{.Rmd} files). If \code{autospin} is set to false (the default),
+#' \code{.R} files will not be spinned nor rendered. \code{autospin} can also
+#' enumerate a list of .R files to be spinned and rendered.
+#'
 #' @section Custom Site Generation:
 #' The behavior of the default site generation function
 #' (\code{rmarkdown::default_site}) is described above. It is also possible to
@@ -326,9 +331,25 @@ default_site <- function(input, encoding = "UTF-8", ...) {
   # helper function to get all input files. includes all .Rmd and
   # .md files that don't start with "_" (note that we don't do this
   # recursively because rmarkdown in general handles applying common
-  # options/elements across subdirectories poorly)
+  # options/elements across subdirectories poorly).
+  # If config$autospin is set to TRUE, then we spin and render .R files as well.
   input_files <- function() {
-    list.files(input, pattern = "^[^_].*\\.[Rr]?md$")
+    if (isTRUE(config$autospin)) {
+      pattern <- "^[^_].*\\.[Rr]?(md)?$"
+    } else {
+      pattern <- "^[^_].*\\.[Rr]?md$"
+    }
+    in_files <- list.files(input, pattern = pattern)
+    if (is.character(config$autospin)) {
+      autospin_src <- file.path(input, config$autospin)
+      missing_autospin_src <- autospin_src[!file.exists(autospin_src)]
+      if (length(missing_autospin_src) > 0) {
+        stop("Missing files in _site.yml: ",
+             paste0(missing_autospin_src, collapse = ", "))
+      }
+      in_files <- c(in_files, config$autospin)
+    }
+    in_files
   }
 
   # define render function (use ... to gracefully handle future args)