+++ tests/testthat/test-site.R
@@ -8,7 +8,7 @@ test_that("render_site", {
   site_dir <- tempfile()
   dir.create(site_dir)
   files <- c("_site.yml", "index.Rmd", "PageA.Rmd",
-             "PageB.rmd", "PageC.md", "styles.css",
+             "PageB.rmd", "PageC.md", "PageD.R", "styles.css",
              "script.R", "docs.txt")
   file.copy(file.path("site", files), site_dir, recursive = TRUE)
 
@@ -16,7 +16,8 @@ test_that("render_site", {
   capture.output(render_site(site_dir))
 
   # did the html files get rendered and the css get copied?
-  html_files <- c("index.html", "PageA.html", "PageB.html", "PageC.html")
+  html_files <- c("index.html", "PageA.html", "PageB.html", "PageC.html",
+                  "PageD.html")
   html_files <- file.path(site_dir, "_site", html_files)
   expect_true(all(file.exists(html_files)))
 
@@ -42,7 +43,7 @@ test_that("render_site respects 'new_session' in the config", {
   # copy parts of our demo site to a tempdir
   site_dir <- tempfile()
   dir.create(site_dir)
-  files <- c("_site.yml", "index.Rmd", "PageA.Rmd", "PageB.rmd")
+  files <- c("_site.yml", "index.Rmd", "PageA.Rmd", "PageB.rmd", "PageD.R")
   file.copy(file.path("site", files), site_dir, recursive = TRUE)
 
   # default behaviour --> new_session: false